+++ tests/testthat/test-knitr-hooks.R
@@ -1,3 +1,31 @@
+test_that("Error thrown: has -check chunk but missing exercise.checker", {
+  rmd <- test_path("tutorials", "missing-exercise-checker.Rmd")
+
+  withr::with_tempfile("outfile", fileext = ".html", {
+    expect_error(
+      rmarkdown::render(rmd, output_file = outfile, quiet = TRUE),
+      regexp = "exercise checker function is not configured"
+    )
+  })
+})
+
+test_that("*-error-check chunks require *-check chunks", {
+  skip_if_not(rmarkdown::pandoc_available())
+
+  tmpfile <- tempfile(fileext = ".html")
+  on.exit(unlink(tmpfile))
+
+  expect_error(
+    rmarkdown::render(test_path("setup-chunks", "error-check-chunk_bad.Rmd"), output_file = tmpfile, quiet = TRUE),
+    "ex-check",
+    fixed = TRUE
+  )
+
+  expect_silent(
+    rmarkdown::render(test_path("setup-chunks", "error-check-chunk_good.Rmd"), output_file = tmpfile, quiet = TRUE)
+  )
+})
+
 test_that("Detection of chained setup cycle works", {
   skip_if_not(rmarkdown::pandoc_available())